new update create new endpoints login and register
Register:
Accepts fullname, email, and password. It validates the inputs, hashes the password, and then sends the data to the User Service to create a new user.

Login:
Accepts email and password. It validates the inputs, then sends a request to the User Service to retrieve user information including email, hashed password, role, and fullname.
After successful verification, it generates a token containing the email, fullname, role, token creation time, and expiration time.
The token is valid for 10 hours.